Role activity module.
Allows monitoring selected roles for selected events. On a busy site with a large
number of users in a given role, provides an alternative to the clutter and volume
in watchdog, which is purged more often than role activity.

Roles to be monitored is configurable. Events are regexps inside the module for
the time being.

Requires the hook_watchdog() patch for Drupal 5.x. See README.txt for details.