Issues #1034612, #1034634 and #1034648: Overhaul the Git gateway management system.
Administrators can now selectively disable Git access. The Git username
(derived from CVS usernames for existing CVS account holders) is now
implemented, and is unchangeable. Considerable feedback is provided
explaining the steps one must take in order to enable Git access. And,
as an added bonus, once the ToS are agreed to and a username is chosen,
instructions appear on how to set up your local git config with the
proper Git name and email.