[#1455614] by dww: Added support for drupal-org-core.make files.
This provides drupal.org distribution maintainers a mechanism to
specify 'fancy' things when building core for use in a distribution,
such as applying patches or checking out a specific Git revision. If
this file does not exist, the script uses th the 'core' attribute from
the drupal-org.make file to build a plain copy of core for packaging.