[#1440412] by dww: Packaging script now allows 'core = 7.x' in .make files.
Previously, it forced you to specify an exact release of core (e.g. 7.12).
Although this is still a *very* good idea, we no longer require it. This
is especially an issue since verify-makefile wasn't considering this
an error so you'd only find out when you actually go to create the release.