[#1821390] by dww: Improve the project add/edit UI with vertical tabs.
If the field_group module is enabled on a site, all the issue-specific
settings on the project node forms will be in an "Issues" vertical tab,
and all the release-specific ones will be grouped under "Releases".
This code iterates over all of the d.o-specific node types to provide
field_group definitions for each bundle.